ai_coordinator/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
google-cloud-aiplatform==1.35.0
pydantic==2.4.2
python-dotenv==1.0.0
